# GestionDuStock_UI/GestionDuStockUI/src/openapi/openapi.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/fournisseurs/{slug}":{"get":{"tags":["supplier-controller"],"operationId":"getFournisseur","parameters":[{"name":"slug","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/supplier"}}}}}},"put":{"tags":["supplier-controller"],"operationId":"updateFournisseur","parameters":[{"name":"slug","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/supplier"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/supplier"}}}}}},"delete":{"tags":["supplier-controller"],"operationId":"deleteFournisseur","parameters":[{"name":"slug","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/api/sales/{saleId}":{"put":{"tags":["sale-controller"],"operationId":"updateSale","parameters":[{"name":"saleId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/saleDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/sale"}}}}}}},"/fournisseurs":{"post":{"tags":["supplier-controller"],"operationId":"createFournisseur","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/supplier"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/supplier"}}}}}}},"/api/sales":{"post":{"tags":["sale-controller"],"operationId":"createSale","parameters":[{"name":"customerId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/saleDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/sale"}}}}}}},"/api/sales/{saleId}/items":{"post":{"tags":["sale-controller"],"operationId":"addSaleItem","parameters":[{"name":"saleId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"productId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"quantity","in":"query","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"price","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/purchases":{"get":{"tags":["purchase-controller"],"operationId":"getAllPurchases","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/purchase"}}}}}}},"post":{"tags":["purchase-controller"],"operationId":"createPurchase","parameters":[{"name":"purchase","in":"query","required":true,"schema":{"$ref":"#/components/schemas/purchase"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"integer","format":"int64"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/purchase"}}}}}}},"/api/products":{"get":{"tags":["product-controller"],"operationId":"getAllProducts","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/product"}}}}}}},"post":{"tags":["product-controller"],"operationId":"createProduct","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/product"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/product"}}}}}}},"/api/customers":{"get":{"tags":["customer-controller"],"operationId":"getAllCustomers","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/customer"}}}}}}},"post":{"tags":["customer-controller"],"operationId":"saveCustomer","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/customer"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/customer"}}}}}}},"/api/auth/register":{"post":{"tags":["auth-controller"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/registerRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/authResponse"}}}}}}},"/api/auth/refresh":{"post":{"tags":["auth-controller"],"operationId":"refresh","responses":{"200":{"description":"OK"}}}},"/api/auth/auth":{"post":{"tags":["auth-controller"],"operationId":"authentication","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/authRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/authResponse"}}}}}}},"/api/sales/invoice/{invoiceNumber}":{"get":{"tags":["sale-controller"],"operationId":"getSaleByInvoiceNumber","parameters":[{"name":"invoiceNumber","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/sale"}}}}}}},"/api/purchases/{id}":{"get":{"tags":["purchase-controller"],"operationId":"getPurchaseBySupplier","parameters":[],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/purchase"}}}}}},"delete":{"tags":["purchase-controller"],"operationId":"deletePurchase","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/purchases/invoice/{invoiceNumber}":{"get":{"tags":["purchase-controller"],"operationId":"getPurchaseByInvoiceNumber","parameters":[{"name":"invoiceNumber","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/purchase"}}}}}}},"/api/products/{id}":{"get":{"tags":["product-controller"],"operationId":"getProductById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/product"}}}}}},"delete":{"tags":["product-controller"],"operationId":"deleteProduct","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/customers/phone/{phone}":{"get":{"tags":["customer-controller"],"operationId":"getCustomerByPhone","parameters":[{"name":"phone","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/customer"}}}}}}},"/api/customers/email/{email}":{"get":{"tags":["customer-controller"],"operationId":"getCustomerByEmail","parameters":[{"name":"email","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/customer"}}}}}}},"/api/sales/{saleId}/items/{saleItemId}":{"delete":{"tags":["sale-controller"],"operationId":"removeSaleItem","parameters":[{"name":"saleId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"saleItemId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/customers/{id}":{"delete":{"tags":["customer-controller"],"operationId":"deleteCustomer","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"product":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"price":{"type":"number","format":"double"},"supplier":{"$ref":"#/components/schemas/supplier"},"stockQuantity":{"type":"integer","format":"int64"},"stockThreshold":{"type":"integer","format":"int64"}}},"supplier":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"slug":{"type":"string"},"nom":{"type":"string"},"email":{"type":"string"},"telephone":{"type":"string"},"adresse":{"type":"string"},"contactPerson":{"type":"string"},"products":{"type":"array","items":{"$ref":"#/components/schemas/product"}},"deleted":{"type":"boolean"}}},"saleDTO":{"type":"object","properties":{"saleItems":{"type":"array","items":{"$ref":"#/components/schemas/saleitemDTO"}},"amount":{"type":"integer","format":"int64"},"date":{"type":"string","format":"date-time"}}},"saleitemDTO":{"type":"object","properties":{"productId":{"type":"integer","format":"int64"},"quantity":{"type":"integer","format":"int64"},"price":{"type":"integer","format":"int64"}}},"customer":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"customerName":{"type":"string"},"email":{"type":"string"},"phone":{"type":"string"},"address":{"type":"string"},"sales":{"type":"array","items":{"$ref":"#/components/schemas/sale"}}}},"sale":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"saleDate":{"type":"string","format":"date-time"},"amount":{"type":"integer","format":"int64"},"invoiceNumber":{"type":"string"},"customer":{"$ref":"#/components/schemas/customer"},"saleItems":{"type":"array","items":{"$ref":"#/components/schemas/saleItem"}}}},"saleItem":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"product":{"$ref":"#/components/schemas/product"},"quantity":{"type":"integer","format":"int64"},"price":{"type":"integer","format":"int64"},"sale":{"$ref":"#/components/schemas/sale"}}},"purchase":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"supplier":{"$ref":"#/components/schemas/supplier"},"products":{"type":"array","items":{"$ref":"#/components/schemas/product"}},"quantity":{"type":"integer","format":"int32"},"totalAmount":{"type":"number","format":"double"},"purchaseDate":{"type":"string","format":"date-time"},"invoiceNumber":{"type":"string"},"autoOrder":{"type":"boolean"},"approved":{"type":"boolean"},"status":{"type":"string"}}},"registerRequest":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"},"role":{"type":"string","enum":["ADMIN","USER"]},"username":{"type":"string"}}},"authResponse":{"type":"object","properties":{"accessToken":{"type":"string"},"refreshToken":{"type":"string"}}},"authRequest":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}}}}}